Refactoring of match4healthcare to curaSWISS
