+++ .gitignore
@@ -8,7 +8,7 @@
 database/data/
 
 # Generated staticfiles by Django
-backend/match4healthcare/static
+backend/curaSWISS/static
 
 # Created by https://www.gitignore.io/api/react,django,python
 # Edit at https://www.gitignore.io/?templates=react,django,python
