+++ backend/apps/accounts/templates/hospital_signup.html
@@ -32,7 +32,7 @@ Registrierung - Institution
 
 
     <h1 class="font-weight-light match4hc-title"> {% blocktrans %}Sie brauchen in Ihrer Institution aktuell Hilfe?{% endblocktrans %} </h1>
-	<h2 class="font-weight-light match4hc-title">{% blocktrans %}Wir verbinden Helfende und Hilfesuchende zu einem match4healthcare!{% endblocktrans %}</h2>
+	<h2 class="font-weight-light match4hc-title">{% blocktrans %}Wir verbinden Helfende und Hilfesuchende zu einem curaSWISS!{% endblocktrans %}</h2>
 
     
             <p>{% blocktrans %}Bitte geben Sie hier die allgemeinen Angaben zu Ihrer Institution an. Nach Klick auf "Registrieren" erhalten Sie eine E-Mail mit einem Bestätigungslink. Nach Klick darauf gelangen Sie in die Datenbank. Dort können nach den von Ihnen gewünschten Kriterien (Ausbildungsstand etc.) Helfende suchen, filtern und diesen auf unserer Plattform direkt eine E-Mail verfassen. Der Emailversand erfolgt dann nach einer Validierung durch uns.{% endblocktrans %}</p>
