+++ backend/apps/accounts/templates/registration/logout.html
@@ -58,7 +58,7 @@ Logout - Studierende
     {% blocktrans %}<p>Danke, dass Du heute etwas Zeit mit der Website verbracht hast.</p>
     <p>Dein Team von</p>
     {% endblocktrans %}
-		<img src="{% static 'img/match4healthcare-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
+		<img src="{% static 'img/curaSWISS-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
 
 		<a href="/accounts/login_redirect"><button type="submit" class="btn blue text-white btn-md" style="margin-top:35px;">{% blocktrans %} Neu anmelden{% endblocktrans %}</button></a>
 </div>
