+++ backend/apps/accounts/templates/registration/password_reset_email_.html
@@ -1,7 +1,7 @@
 {% load i18n %}
 
 {% blocktrans %}
-Sie erhalten diese Mail um ein neues Passwort bei match4healthcare zu setzen.
+Sie erhalten diese Mail um ein neues Passwort bei curaSWISS zu setzen.
 
 Bitte clicken sie auf den folgenden Link um ein Passwort zu wählen:
 {% endblocktrans %}
@@ -12,5 +12,5 @@ Bitte clicken sie auf den folgenden Link um ein Passwort zu wählen:
 
 {% blocktrans %}
 Vielen Dank,
-ihr match4healthcare Team
+ihr curaSWISS Team
 {% endblocktrans %}
