+++ backend/apps/accounts/templates/registration/password_reset_email_subject.txt
@@ -1,2 +1,2 @@
 {% load i18n %}
-[match4healthcare] {% blocktrans %}Email bestätigen{% endblocktrans %}
+[curaSWISS] {% blocktrans %}Email bestätigen{% endblocktrans %}
