+++ backend/apps/accounts/templates/registration/password_set_email_hospital.html
@@ -14,7 +14,7 @@
 		<meta charset="UTF-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>match4healthcare</title>
+		<title>curaSWISS</title>
         
     <style type="text/css">
 		p{
@@ -397,7 +397,7 @@
                             <td class="mcnImageContent" valign="top" style="padding-right: 9px;padding-left: 9px;padding-top: 0;padding-bottom: 0;text-align: center;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;">
                                 
                                     
-                                        <img align="left" alt="" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/match4healthcare-logo.png" width="250" style="max-width: 250px;padding-bottom: 0;display: inline !important;vertical-align: bottom;border: 0;height: auto;outline: none;text-decoration: none;-ms-interpolation-mode: bicubic;" class="mcnImage">
+                                        <img align="left" alt="" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/curaSWISS-logo.png" width="250" style="max-width: 250px;padding-bottom: 0;display: inline !important;vertical-align: bottom;border: 0;height: auto;outline: none;text-decoration: none;-ms-interpolation-mode: bicubic;" class="mcnImage">
                                     
                                 
                             </td>
@@ -444,7 +444,7 @@
                         
                             <div style="text-align: left;"><span style="font-size:18px; font-weight: 500;"><span style="color:#0d2333">{% blocktrans %}Sehr geehrte Einrichtung!{% endblocktrans %}</span></span><br>
                               
-								<span>{% blocktrans %}Vielen Dank für Ihren Eintrag auf match4healthcare!{% endblocktrans %}</span>
+								<span>{% blocktrans %}Vielen Dank für Ihren Eintrag auf curaSWISS!{% endblocktrans %}</span>
                               <span style="font-size:15px; color: #535353;"> 
 <br><br>
 {% blocktrans %}Bitte bestätigen Sie jetzt Ihre Registrierung, indem Sie auf den folgenden Link klicken:{% endblocktrans %}
@@ -468,9 +468,9 @@
                               <br><span style="font-size:15px; color: #535353;">
 								
 								In jedem Fall ein großes Dankeschön für das große Engagement Ihrer Einrichtung in der aktuellen Lage. Wir hoffen, mit dieser Plattform einen kleinen Beitrag für die aktuelle Gesundheitsversorgung geleistet zu haben.
-Für Anregungen zu unserer Plattform und bei weiteren Fragen kontaktieren Sie uns gerne unter <a href="mailto:info@match4healthcare.de" style="color:#ed0a71; text-decoration:none;">info@match4healthcare.de</a>.
+Für Anregungen zu unserer Plattform und bei weiteren Fragen kontaktieren Sie uns gerne unter <a href="mailto:info@curaSWISS.de" style="color:#ed0a71; text-decoration:none;">info@curaSWISS.de</a>.
 </span><br><br>
-                         <span style="font-size:15px; color: #535353;">Liebe Grüße,<br>Ihr match4healthcare Team</span>
+                         <span style="font-size:15px; color: #535353;">Liebe Grüße,<br>Ihr curaSWISS Team</span>
                         <br><br>
                            
                            
@@ -481,12 +481,12 @@ Für Anregungen zu unserer Plattform und bei weiteren Fragen kontaktieren Sie un
    
 
       <p style="font-family: Helvetica, Arial, sans-serif; font-size: 15px; line-height: 150%; margin-bottom: 15px;">
-        <span style="font-weight: bold; color: #535353; display: inline;">match4healthcare</span>
+        <span style="font-weight: bold; color: #535353; display: inline;">curaSWISS</span>
         <span style="display: inline;"><br></span>
 
         
       
-        <a href="https://www.future-doctor.de" style="color: #ed0a71; text-decoration: none; display: inline;">www.match4healthcare.de</a>
+        <a href="https://www.future-doctor.de" style="color: #ed0a71; text-decoration: none; display: inline;">www.curaSWISS.de</a>
       </p>
 
 
@@ -494,26 +494,26 @@ Für Anregungen zu unserer Plattform und bei weiteren Fragen kontaktieren Sie un
     
     
     <a style="text-decoration: none; display: inline;" href="https://twitter.com/match4health_">
-      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="facebook.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/twitter.png" alt="Facebook">
+      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="facebook.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/twitter.png" alt="Facebook">
     </a>
     <span style="white-space: nowrap; display: inline;">
-      <img src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/spacer.gif" width="10">
+      <img src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/spacer.gif" width="10">
     </span>
 
     
     
     <a style="text-decoration: none; display: inline;" href="https://www.youtube.com/watch?v=zZwdMESJav0&feature=emb_title">
-      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="youtube.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/youtube.png" alt="Youtube">
+      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="youtube.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/youtube.png" alt="Youtube">
     </a>
 		
 		<span style="white-space: nowrap; display: inline;">
-      <img src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/spacer.gif" width="10">
+      <img src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/spacer.gif" width="10">
     </span>
 
     
     
-    <a style="text-decoration: none; display: inline;" href="https://www.instagram.com/match4healthcare/">
-      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="youtube.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-match4healthcare/instagram.png" alt="Youtube">
+    <a style="text-decoration: none; display: inline;" href="https://www.instagram.com/curaSWISS/">
+      <img width="35" style="margin-bottom:2px; border:none; display:inline;" height="35" data-filename="youtube.png" src="https://medis-vs-covid19.de/wp-content/uploads/signatur-curaSWISS/instagram.png" alt="Youtube">
     </a>
