+++ backend/apps/iamstudent/templates/thanks.html
@@ -53,11 +53,11 @@
   <path fill-rule="evenodd" d="M8 2.5A5.5 5.5 0 1013.5 8a.5.5 0 011 0 6.5 6.5 0 11-3.25-5.63.5.5 0 11-.5.865A5.472 5.472 0 008 2.5z" clip-rule="evenodd"/>
 </svg>
 
-    {% blocktrans %}<h2 class="font-weight-light">Willkommen bei match4healthcare,</h2>
+    {% blocktrans %}<h2 class="font-weight-light">Willkommen bei curaSWISS,</h2>
     <p>Um deine Registrierung zu vervollständigen, haben wir dir eine Mail mit einem <b style="font-weight:500;">Bestätigungslink</b> an deine angegebene Mailadresse geschickt. Bitte schau in deine Mails, um dich vollständig bei uns zu registrieren. <br><br><b>Aktuell kann es zu Verzögerungen beim Eingang der E-Mail kommen. Wir bitten um Geduld und dein Verständnis.</b></p>
     <p>Dein Team von</p>
     {% endblocktrans %}
-		<img src="{% static 'img/match4healthcare-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
+		<img src="{% static 'img/curaSWISS-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
 </div>
 
 	</div>
