+++ backend/apps/ineedstudent/templates/thanks_hospital.html
@@ -27,11 +27,11 @@
   <path fill-rule="evenodd" d="M8 2.5A5.5 5.5 0 1013.5 8a.5.5 0 011 0 6.5 6.5 0 11-3.25-5.63.5.5 0 11-.5.865A5.472 5.472 0 008 2.5z" clip-rule="evenodd"/>
 </svg>
 	
-    {% blocktrans %}<h2 class="font-weight-light">Willkommen bei match4healthcare,</h2>
+    {% blocktrans %}<h2 class="font-weight-light">Willkommen bei curaSWISS,</h2>
     <p>vielen Dank, dass Sie sich auf unserer Plattform registriert haben. Auf der nachfolgenden Seite haben Sie sich Möglichkeit, sich einen Überblick über die verfügbaren Helfer zu verschaffen.</p>
     <p>Dein Team von</p>
     {% endblocktrans %}
-		<img src="{% static 'img/match4healthcare-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
+		<img src="{% static 'img/curaSWISS-logo-text.svg' %}" width="180px" style="vertical-align: baseline">
 		<a href="#"><button type="submit" class="btn blue text-white btn-md" style="margin-top:35px;">{% blocktrans %} HelferInnen anzeigen{% endblocktrans %}</button></a>
 </div>
