+++ backend/curaSWISS/settings/common.py
@@ -1,5 +1,5 @@
 """
-Django settings for Match4healthcare project.
+Django settings for curaSWISS project.
 
 Generated by 'django-admin startproject' using Django 3.0.4.
 
@@ -48,7 +48,7 @@ MIDDLEWARE = [
     
 ]
 
-ROOT_URLCONF = 'match4healthcare.urls'
+ROOT_URLCONF = 'curaSWISS.urls'
 
 TEMPLATES = [
     {
@@ -70,9 +70,9 @@ LOGIN_REDIRECT_URL = '/accounts/login_redirect'
 
 CRISPY_TEMPLATE_PACK = 'bootstrap4'
 
-NOREPLY_MAIL = 'match4healthcare<noreply@match4healthcare.de>'
+NOREPLY_MAIL = 'curaSWISS<noreply@curaSWISS.de>'
 
-WSGI_APPLICATION = 'match4healthcare.wsgi.application'
+WSGI_APPLICATION = 'curaSWISS.wsgi.application'
 
 MAX_EMAIL_BATCH_PER_HOSPITAL = 200
