+++ backend/curaSWISS/settings/production.py
@@ -1,4 +1,4 @@
-from match4healthcare.settings.common import *
+from curaSWISS.settings.common import *
 from django.utils.log import DEFAULT_LOGGING
 
 import logging
@@ -11,8 +11,8 @@ DEBUG = False
 
 SECRET_KEY = os.environ['SECRET_KEY']
 
-ALLOWED_HOSTS = ['matchmedisvsvirus.dynalias.org', 'helping-health.from-de.com', 'match4healthcare.de',
-                 'match4healthcare.eu', 'match4healthcare.org', 'medis-vs-covid19.de']
+ALLOWED_HOSTS = ['matchmedisvsvirus.dynalias.org', 'helping-health.from-de.com', 'curaSWISS.de',
+                 'curaSWISS.eu', 'curaSWISS.org', 'medis-vs-covid19.de']
 
 # Database
 # https://docs.djangoproject.com/en/3.0/ref/settings/#databases
@@ -44,7 +44,7 @@ LOGGING = {
         'applogfile': {
             'level': 'ERROR',
             'class': 'logging.handlers.RotatingFileHandler',
-            'filename': os.path.join(BASE_DIR, 'run', 'match4healthcare.log'),
+            'filename': os.path.join(BASE_DIR, 'run', 'curaSWISS.log'),
             'maxBytes': 1024 * 1024 * 15,  # 15MB
             'backupCount': 10,
         },
