+++ backend/curaSWISS/views.py
