+++ backend/templates/about.html
@@ -10,7 +10,7 @@
     <div class="container">
         <div style="height:40px"></div>
         <h1 style="text-align:center; font-weight: 300;">{% blocktrans %}Wir sind das Team von
-            match4healthcare!{% endblocktrans %}</h1>
+            curaSWISS!{% endblocktrans %}</h1>
         <hr>
         <p class="text-center">{% blocktrans %}Durch die bundesweit erfolgreich und bekannt gewordenen Initiative
             „Medizinstudierende vs. COVID19“ (<a href="https://www.medis-vs-covid19.de" target="_blank">medis-vs-covid19.de</a>), die Bundesvertretung der
