+++ backend/templates/base.html
@@ -33,19 +33,19 @@
 <meta name="msapplication-TileImage" content="/ms-icon-144x144.png">
 <meta name="theme-color" content="#ffffff">
 <meta property="og:type" content="website">
-<meta property="og:image" content="https://match4healthcare.de{% static 'img/facebook-share.jpg' %}">
-<meta property="og:image:secure_url" content="https://match4healthcare.de{% static 'img/facebook-share.jpg' %}">
-<meta name="twitter:image" content="https://match4healthcare.de{% static 'img/facebook-share.jpg' %}">
+<meta property="og:image" content="https://curaSWISS.de{% static 'img/facebook-share.jpg' %}">
+<meta property="og:image:secure_url" content="https://curaSWISS.de{% static 'img/facebook-share.jpg' %}">
+<meta name="twitter:image" content="https://curaSWISS.de{% static 'img/facebook-share.jpg' %}">
 <meta property="og:image:width" content="1200">
 <meta property="og:image:height" content="630">
-<title>{% block title %} {% endblock %} | match4healthcare </title>
+<title>{% block title %} {% endblock %} | curaSWISS </title>
 {% block header %}
     {% endblock %}
 </head>
 
 <body>
 <nav class="navbar navbar-expand-md fixed-top navbar-light bg-white">
-  <div class="container"> <a class="navbar-brand" href="/"><img src="{% static 'img/match4healthcare-logo.svg' %}" width="199px"></a>
+  <div class="container"> <a class="navbar-brand" href="/"><img src="{% static 'img/curaSWISS-logo.svg' %}" width="199px"></a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExampleDefault"
                 aria-controls="navbarsExampleDefault" aria-expanded="false" aria-label="Toggle navigation"> <span class="navbar-toggler-icon"></span> </button>
     <div class="navbar-collapse collapse w-100 order-1 order-md-0 dual-collapse2" id="navbarsExampleDefault">
@@ -80,7 +80,7 @@
   {% endblock %} </div>
 <footer class="footer">
   <div class="container">
-    <p style="margin-bottom:5px !important;"> <a style="text-decoration: none; display: inline;" href="https://twitter.com/match4health_" target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/twitter.svg' %}" alt="Twitter"> </a> <a style="text-decoration: none; display: inline;" href="https://www.instagram.com/match4healthcare/"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; margin-left:8px; display:inline;" height="25" src="{% static 'img/instagram.svg' %}" alt="Instagram"> </a> <a style="text-decoration: none; display: inline; margin-left:8px;" href="https://www.facebook.com/Match4Healthcare/"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/facebook.svg' %}" alt="Facebook"> </a> <a style="text-decoration: none; display: inline; margin-left:8px;" href="https://www.youtube.com/watch?v=zZwdMESJav0&feature=emb_title"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/youtube.svg' %}" alt="Youtube"> </a></p>
+    <p style="margin-bottom:5px !important;"> <a style="text-decoration: none; display: inline;" href="https://twitter.com/match4health_" target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/twitter.svg' %}" alt="Twitter"> </a> <a style="text-decoration: none; display: inline;" href="https://www.instagram.com/curaSWISS/"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; margin-left:8px; display:inline;" height="25" src="{% static 'img/instagram.svg' %}" alt="Instagram"> </a> <a style="text-decoration: none; display: inline; margin-left:8px;" href="https://www.facebook.com/curaSWISS/"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/facebook.svg' %}" alt="Facebook"> </a> <a style="text-decoration: none; display: inline; margin-left:8px;" href="https://www.youtube.com/watch?v=zZwdMESJav0&feature=emb_title"  target="_blank"> <img width="25" style="margin-bottom:0px; border:none; display:inline;" height="25" src="{% static 'img/youtube.svg' %}" alt="Youtube"> </a></p>
     <p>{% blocktrans %}Ein Projekt im Rahmen des{% endblocktrans %} <a href="https://wirvsvirushackathon.org/"
                     target="_blank">#WirVsVirus Hackathons</a>. <a
                     href="/impressum">{% blocktrans %}Impressum{% endblocktrans %}</a> {% blocktrans %}und{% endblocktrans %} <a
