+++ backend/templates/dataprotection.html
@@ -25,7 +25,7 @@ Datenschutzerklärung
         <div class="row">
             <div class="col-md-4 m-1 p-3">
                 <h5>Wer wir sind / Kontakt­information</h5>
-                <strong>Initiative match4healthcare</strong>
+                <strong>Initiative curaSWISS</strong>
                <br>
                 Tel.: +49 7121 360 5434 <br>
                 E-Mail: <script type="text/javascript">document.write(decode("vasb@zrqvf-if-pbivq19.qr"))</script><br><br>
@@ -57,7 +57,7 @@ Datenschutzerklärung
                 </p>
 
                 <h5>Löschung der Daten:</h5>
-                <p>Sollte eine Löschung der Daten gewünscht sein, kann dies per email an datenschutz@match4healthcare.de angezeigt werden.
+                <p>Sollte eine Löschung der Daten gewünscht sein, kann dies per email an datenschutz@curaSWISS.de angezeigt werden.
                     Rechtsgrundlage:<br>
                     Sie haben grundsätzlich folgende Rechte:<br>
                     Auskunftsrecht (Art. 15 DSGVO)<br>
@@ -101,7 +101,7 @@ Datenschutzerklärung
                     Quelle: Teile dieser Datenschutzerklärung wurden mit dem Datenschutz Generator von firmenwebseiten.at erstellt.<br>
                 </p>
                 <h5>Haftungsausschluss AGB/Datenschutz</h5>
-                <p>Wir verbinden Helfende mit Hilfesuchenden. Der Einsatz erfolgt in Eigenverantwortung und in Verantwortung der entsprechenden Einrichtung. match4healthcare übernimmt weder die Haftung noch etwaige notwendige Versicherungen für die Einsätze, die im Anschluss an ein matching durch match4healthcare entstehen. Ebenfalls erfolgt die Angabe der Informationen im Portal auf Eigenverantwortung. Qualifikationen und alle angegebenen Informationen der NutzerInnen müssen vor Einsatzbeginn von Seiten der Hilfesuchenden und Helfenden nochmals verifiziert werden.</p>
+                <p>Wir verbinden Helfende mit Hilfesuchenden. Der Einsatz erfolgt in Eigenverantwortung und in Verantwortung der entsprechenden Einrichtung. curaSWISS übernimmt weder die Haftung noch etwaige notwendige Versicherungen für die Einsätze, die im Anschluss an ein matching durch curaSWISS entstehen. Ebenfalls erfolgt die Angabe der Informationen im Portal auf Eigenverantwortung. Qualifikationen und alle angegebenen Informationen der NutzerInnen müssen vor Einsatzbeginn von Seiten der Hilfesuchenden und Helfenden nochmals verifiziert werden.</p>
             </div>
         </div>
