+++ backend/templates/home.html
@@ -8,13 +8,13 @@ Home
 {% block header %}
 <meta name="description" content="Der Glaube daran Mitmenschen in Not helfen können, die Liebe zum Fach Medizin und der Zusammenhalt untereinander sind in den letzten Wochen und Monaten im Gesundheitswesen noch viel präsenter denn je.">
 <meta name="robots" content="max-snippet:-1, max-image-preview:large, max-video-preview:-1">
-<link rel="canonical" href="https://match4healthcare.de/">
-<meta property="og:title" content="match4healthcare - Wir vernetzen Helfende mit Hilfesuchenden.">
+<link rel="canonical" href="https://curaSWISS.de/">
+<meta property="og:title" content="curaSWISS - Wir vernetzen Helfende mit Hilfesuchenden.">
 <meta property="og:description" content="Der Glaube daran Mitmenschen in Not helfen können, die Liebe zum Fach Medizin und der Zusammenhalt untereinander sind in den letzten Wochen und Monaten im Gesundheitswesen noch viel präsenter denn je.">
-<meta property="og:url" content="https://match4healthcare.de/">
-<meta property="og:site_name" content="match4healthcare">
+<meta property="og:url" content="https://curaSWISS.de/">
+<meta property="og:site_name" content="curaSWISS">
 <meta name="twitter:description" content="Der Glaube daran Mitmenschen in Not helfen können, die Liebe zum Fach Medizin und der Zusammenhalt untereinander sind in den letzten Wochen und Monaten im Gesundheitswesen noch viel präsenter denn je.">
-<meta name="twitter:title" content="match4healthcare - Wir vernetzen Helfende mit Hilfesuchenden.">
+<meta name="twitter:title" content="curaSWISS - Wir vernetzen Helfende mit Hilfesuchenden.">
 <style>
 	body {
   margin-top: 0px;
@@ -90,7 +90,7 @@ Home
   <div class="home-title">
     <div class="container">
       <div class="row align-self-center align-items-center">
-        <div class="col-md-8 col-xl-5"> <img src="{% static 'img/match4healthcare-logo.svg' %}" width="320px" class="logo-header">
+        <div class="col-md-8 col-xl-5"> <img src="{% static 'img/curaSWISS-logo.svg' %}" width="320px" class="logo-header">
           <h1 class="text-white font-weight-light">{% blocktrans %}In Zeiten der Krise vernetzen wir Helfende mit Hilfesuchenden.{% endblocktrans %}</h1>
         </div>
       </div>
@@ -132,13 +132,13 @@ Home
   <div class="container" style="margin-top:55px;">
     <div class="row">
       <div class="col-md-12" style="text-align: left; margin: 0 auto;">
-        <h2 style="text-align:center; font-size: 1.5rem; font-weight:600; margin-bottom: 15px">{% blocktrans %}Gemeinsam unser Gesundheitssystem stärken, mit einem{% endblocktrans %} <img src="{% static 'img/match4healthcare-logo-text.svg' %}" width="180px" style="vertical-align: baseline"></h2>
+        <h2 style="text-align:center; font-size: 1.5rem; font-weight:600; margin-bottom: 15px">{% blocktrans %}Gemeinsam unser Gesundheitssystem stärken, mit einem{% endblocktrans %} <img src="{% static 'img/curaSWISS-logo-text.svg' %}" width="180px" style="vertical-align: baseline"></h2>
         <hr style="width:60%; margin: 0 auto; margin-bottom:45px; border-width: 2px; ">
       </div>
       <div class="col-lg-7 order-sm-2 order-lg-3" style="text-align: left; margin-bottom: 55px;">
         <p>{% blocktrans %}Der Glaube daran, dass Mitmenschen in Not helfen können, die Liebe zum Fach Medizin und der Zusammenhalt untereinander sind in den letzten Wochen und Monaten im Gesundheitswesen noch viel präsenter denn je.{% endblocktrans %}</p>
         <p>{% blocktrans %}Gerade aus diesem Grund brauchen wir zur jetzigen Zeit eine einfache und vor allen Dingen eine effiziente Schnittstelle, zwischen den medizinischen Einrichtungen und Fachkräften. Eine Schnittstelle, die es ermöglicht, dass Hilfebietende und Hilfesuchende schnell und unkompliziert miteinander vernetzt werden.{% endblocktrans %}</p>
-        <p>{% blocktrans %}Die Plattform "match4healthcare" bietet die ideale Schnittstelle zur schnellen und effizienten Vernetzung eben dieser Parteien und bietet vor allem zu Zeiten wie jetzt die ideale Problemlösung.{% endblocktrans %}</p>
+        <p>{% blocktrans %}Die Plattform "curaSWISS" bietet die ideale Schnittstelle zur schnellen und effizienten Vernetzung eben dieser Parteien und bietet vor allem zu Zeiten wie jetzt die ideale Problemlösung.{% endblocktrans %}</p>
       </div>
       <div class="col-lg-5 order-sm-3 order-lg-2" style="text-align: center; margin-bottom: 55px;">
         <div class="embed-responsive embed-responsive-16by9">
