+++ backend/templates/legal-questions.html
@@ -12,7 +12,7 @@
         <h2 style="font-weight: 300;">Disclaimer</h2>
         <hr>
         <p >
-            Wir fassen hier nach unserem besten Wissen und Gewissen Tipps und Fragestellungen zusammen, von denen wir denken, dass du sie vor deiner Tätigkeit als Helfende_r während der Corona Krise durchgehen solltest. Wir erheben keinen Anspruch auf die Vollständigkeit und Korrektheit dieser Zusammenfassung. match4healthcare übernimmt weder die Haftung noch etwaige notwendige Versicherungen für die Einsätze, die im Anschluss an ein Matching durch match4healthcare entstehen.
+            Wir fassen hier nach unserem besten Wissen und Gewissen Tipps und Fragestellungen zusammen, von denen wir denken, dass du sie vor deiner Tätigkeit als Helfende_r während der Corona Krise durchgehen solltest. Wir erheben keinen Anspruch auf die Vollständigkeit und Korrektheit dieser Zusammenfassung. curaSWISS übernimmt weder die Haftung noch etwaige notwendige Versicherungen für die Einsätze, die im Anschluss an ein Matching durch curaSWISS entstehen.
         </p>
 
         <h2 style="font-weight: 300;">Auf was müsst ihr beim Unterschreiben eines Vertrags achten?</h2>
