+++ docker-compose.dev.yml
@@ -8,7 +8,7 @@ services:
       context: ./backend
       dockerfile: Dockerfile
     volumes:
-      - './backend:/match4healthcare-backend'
+      - './backend:/curaSWISS-backend'
     ports:
       - '8000:8000'
